+++ rest/src/main/kotlin/com/gitlab/kordlib/rest/ratelimit/RequestHandler.kt
@@ -5,6 +5,6 @@ import io.ktor.client.statement.HttpResponse
 
 interface RequestHandler {
 
-    suspend fun <T> handle(request: Request<T>): HttpResponse
+    suspend fun <B : Any,R>  handle(request: Request<B,R>): R
 
 }