+++ rest/src/main/kotlin/com/gitlab/kordlib/rest/route/Route.kt
@@ -63,6 +63,9 @@ sealed class Route<T>(
     object ReactionDelete
         : Route<Unit>(HttpMethod.Delete, "/channels/$ChannelId/messages/$MessageId/reactions/$Emoji/$UserId", NoStrategy)
 
+    object DeleteAllReactionsForEmoji
+        : Route<Unit>(HttpMethod.Delete, "/channels/$ChannelId/messages/$MessageId/reactions/$Emoji", NoStrategy)
+
     object MessageDelete
         : Route<Unit>(HttpMethod.Delete, "/channels/$ChannelId/messages/$MessageId", NoStrategy)