+++ CHANGELOG.md
@@ -3,9 +3,11 @@
 ## Additions
 
 * `Kord` and its`Cache` now implement a common interface `EntitySupplier` to retrieve entities that can be cached. 
+* `mentionedRoleIds`, `mentionedRoleBehaviors`, `mentionedUserIds`, `mentionedUserBehaviors` were added to `Message`.
 
 ## Changes
 
+* `Message#mentionedRoles` and `Message#mentionedUsers` now return a `Flow` of their respective entities instead of a `Set<Snwoflake>`.
 * `StoreChannel#edit`, `TextChannel#edit` and `NewsChannel#edit` now supply their builder as a receiver.
 * core entity builders were moved from `com.gitlab.kordlib.core.builder` to `com.gitlab.kordlib.rest.builder` and are now
 part of the rest module.
@@ -17,9 +19,18 @@ part of the rest module.
 
 # 0.3.2
 
+## Additions
+
+* Added `InviteCreate`, `InviteDelete` an `MessageReactionRemoveEmoji` events. #61
+* Added `deleteAllReactionsForEmoji` to ChannelService. #61
+
 ## Fixes 
 
 * Fixed an issue where `DiscordInvite` was wrongly representing `inviter` as a `String` instead of a `DiscordUser`.
+
+## Dependencies
+
+* ktor 1.3.0-rc2 -> 1.3.0
 
 # 0.3.1